backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.0
langchain-groq==0.2.0
langchain-community==0.3.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.8.2
sse-starlette==2.1.3
